Move recovery tests to separate file
